% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sgk_get_latest_year}
\alias{sgk_get_latest_year}
\title{Get the most recent available year}
\usage{
sgk_get_latest_year()
}
\value{
Integer with the most recent year available in the SGK API
}
\description{
Get the most recent available year
}
\examples{
\dontrun{
latest_year <- sgk_get_latest_year()
}
}
