% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgkveri-package.R
\docType{package}
\name{sgkveri-package}
\alias{sgkveri-package}
\title{sgkveri: Programmatic downloads from SGK Veri}
\description{
Automates indicator selection and data downloads from net.sgk.gov.tr/sgkveri2 via API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emraher/sgkveri}
  \item Report bugs at \url{https://github.com/emraher/sgkveri/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com}

}
\keyword{internal}
